Backed out changeset e1297b6262c4 (bug 1674353) for causing build bustages in AsmJS. CLOSED TREE
UltraBlame original commit: 2093e9abe864765b8d0507797f2833af3c088ff7

+++ js/src/wasm/AsmJS.h
@@ -106,7 +106,7 @@ extern JSString* AsmJSModuleToString(JSContext* cx, JS::Handle<JSFunction*> fun,
 
 
 
-extern bool IsValidAsmJSHeapLength(size_t length);
+extern bool IsValidAsmJSHeapLength(uint32_t length);
 
 }
